data_construction/requirements.txt
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1691435290924/work
crcmod==1.7
cryptography==42.0.5
cycler==0.10.0
datasets==2.13.0
dill==0.3.6
Django==3.2.20
einops==0.6.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.5.4
h5py==2.9.0
huggingface-hub==0.16.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
image==1.5.27
imageio==2.5.0
importlib-metadata==6.7.0
jmespath==0.10.0
kiwisolver==1.1.0
matplotlib==3.3.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
modelscope==1.9.5
multidict==6.0.5
multiprocess==0.70.14
networkx==2.6.3
numpy==1.21.6
opencv-python==4.8.0.76
oss2==2.18.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.3.5
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.20.0
pygame==2.5.1
pyparsing==2.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
scikit-image==0.19.0
scipy==1.7.3
simplejson==3.19.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sortedcontainers==2.4.0
sqlparse==0.3.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado==6.0.2
tqdm==4.66.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
wget==3.2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.15.0
